requirements.txt
google-adk==1.26.0
google-genai==1.66.0
python-telegram-bot==22.6
requests==2.32.5
qrcode==8.2
Pillow>=10.0.0
Flask==3.1.0
pydantic==2.12.5
google-cloud-storage==2.18.2
